+++ src/createDateObjects.js
@@ -17,7 +17,10 @@ export default function createDateObjects(date, weekOffset = 0) {
   }));
 
   const daysAdded = prevMonthDays.length + currentMonthDays.length - 1;
-  const nextMonthDays = takeWhile(range(1, 7), n => (daysAdded + n) % 7 !== 0).map((n) => ({
+  const nextMonthDays = takeWhile(
+    range(1, 7),
+    n => (daysAdded + n) % 7 !== 0
+  ).map(n => ({
     day: last(currentMonthDays).day.clone().add(n, 'days'),
     classNames: 'nextMonth'
   }));